Retail/requirements.txt
dtlpy
pandas
numpy
Pillow
opencv-python
pathlib 